Day-27/Airflow Demo - ELT/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pyarrow
pandas
sqlalchemy
psycopg2-binary